Move browser common files into xray_web
