+++ xray_web/lib/app.js
