+++ xray_web/lib/modal.js
