+++ xray_web/lib/text_editor/shaders.js
