+++ xray_web/lib/text_editor/text_editor.js
