+++ xray_web/lib/text_editor/text_plane.js
