+++ xray_web/lib/view.js
