+++ xray_web/lib/view_registry.js
