+++ xray_web/lib/workspace.js
