+++ xray_web/package.json
@@ -1,23 +1,21 @@
 {
-  "name": "Xray",
-  "version": "0.1.0",
-  "main": "./lib/main_process/main.js",
+  "name": "xray_web",
   "license": "MIT",
   "scripts": {
-    "start": "electron .",
     "test": "electron-mocha --ui=tdd --renderer test/**/*.test.js",
     "itest": "electron-mocha --ui=tdd --renderer --interactive test/**/*.test.js"
   },
   "dependencies": {
+    "prop-types": "^15.6.0",
     "react": "^16.3.0",
     "react-dom": "^16.3.0",
-    "xray_web": "../xray_web"
+    "styletron-engine-atomic": "^1.0.4",
+    "styletron-react": "^4.0.3"
   },
   "devDependencies": {
     "electron": "2.0.0-beta.7",
     "electron-mocha": "^6.0.1",
     "enzyme": "^3.3.0",
-    "enzyme-adapter-react-16": "^1.1.1",
-    "prettier": "^1.10.2"
+    "enzyme-adapter-react-16": "^1.1.1"
   }
 }
