+++ xray_web/test/helpers/component_helpers.js
